requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
crispy-bootstrap5==2024.2
Django==4.2.14
django-bootstrap-v5==1.0.11
django-cors-headers==4.4.0
django-crispy-forms==2.2
numpy==2.0.0
pandas==2.2.2
pillow==10.4.0
pypdf==4.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
